backend/requirements.txt
django-rest-framework-0.1.0
django-3.1.6
django-rest-knox-0.1.11
PyJWT-1.7.1
djangorestframework-jwt-1.11.0
django-cors-headers-3.3.1